Move the tests
