+++ src-tauri/src/tests/mod.rs
@@ -1,3 +1,4 @@
 // Test modules
 mod tests_parser;
 mod test_replay_parser;
+mod tests_lib;
